Rename: Loader컴포넌트 common폴더 하위로 이동
