--- webapp/src/components/Common/Loader/index.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Loader() {
-  return <div>🧐 Loading...</div>;
-}
-
-export default Loader;
